feat: getting music to read and displaying it in the screen
BREAKING CHANGE : this commit content lot of changes.
Firstly, some functions have been commented and the unused class MusicLinkResponse has been removed.
We then implemented the function getMusicReader in MusicFinderService so that to get the music with
the song url.
Finally, we changed the code of some classes in order to have a beautiful code.